feat: Set `audience` field in authenticated HTTP task example (#138)
* Add explicit `audience1 field

Unset `audience` will be take the URL as default value. When there are query and fragment parts, this would lead to authentication errors (401).

* Add explicit `audience1 field

feat: Set `audience` in authenticated HTTP task

Unset `audience` will be take the URL as default value. When there are query and fragment parts, this would lead to authentication errors (401).

* feat: Add explicit `audience1 field

Unset `audience` will be take the URL as default value. When there are query and fragment parts, this would lead to authentication errors (401).